@@ -29,8 +29,7 @@ var selfEvents = require('cordova-common').events;
 
 var PLATFORM = 'android';
 
-
-function setupEvents(externalEventEmitter) {
+function setupEvents (externalEventEmitter) {
     if (externalEventEmitter) {
         // This will make the platform internal events visible outside
         selfEvents.forwardEventsTo(externalEventEmitter);
@@ -43,7 +42,6 @@ function setupEvents(externalEventEmitter) {
     return selfEvents;
 }
 
-
 /**
  * Class, that acts as abstraction over particular platform. Encapsulates the
  *   platform's properties and methods.
@@ -55,7 +53,7 @@ function setupEvents(externalEventEmitter) {
  *
  * * platform: String that defines a platform name.
  */
-function Api(platform, platformRootDir, events) {
+function Api (platform, platformRootDir, events) {
     this.platform = PLATFORM;
     this.root = path.resolve(__dirname, '..');
 
@@ -79,14 +77,14 @@ function Api(platform, platformRootDir, events) {
     };
 
     // XXX Override some locations for Android Studio projects
-    if(AndroidStudio.isAndroidStudioProject(self.root) === true) {
-      selfEvents.emit('log', 'Android Studio project detected');
-      this.android_studio = true;
-      this.locations.configXml = path.join(self.root, 'app/src/main/res/xml/config.xml');
-      this.locations.strings = path.join(self.root, 'app/src/main/res/xml/strings.xml');
-      this.locations.manifest = path.join(self.root, 'app/src/main/AndroidManifest.xml');
-      this.locations.www = path.join(self.root, 'app/src/main/assets/www');
-      this.locations.res = path.join(self.root, 'app/src/main/res');
+    if (AndroidStudio.isAndroidStudioProject(self.root) === true) {
+        selfEvents.emit('log', 'Android Studio project detected');
+        this.android_studio = true;
+        this.locations.configXml = path.join(self.root, 'app/src/main/res/xml/config.xml');
+        this.locations.strings = path.join(self.root, 'app/src/main/res/xml/strings.xml');
+        this.locations.manifest = path.join(self.root, 'app/src/main/AndroidManifest.xml');
+        this.locations.www = path.join(self.root, 'app/src/main/assets/www');
+        this.locations.res = path.join(self.root, 'app/src/main/res');
     }
 }
 
@@ -112,16 +110,13 @@ Api.createPlatform = function (destination, config, options, events) {
     events = setupEvents(events);
     var result;
     try {
-        result = require('../../lib/create')
-        .create(destination, config, options, events)
-        .then(function (destination) {
+        result = require('../../lib/create').create(destination, config, options, events).then(function (destination) {
             var PlatformApi = require(path.resolve(destination, 'cordova/Api'));
             return new PlatformApi(PLATFORM, destination, events);
         });
-    }
-    catch (e) {
-        events.emit('error','createPlatform is not callable from the android project API.');
-        throw(e);
+    } catch (e) {
+        events.emit('error', 'createPlatform is not callable from the android project API.');
+        throw (e);
     }
     return result;
 };
@@ -146,16 +141,13 @@ Api.updatePlatform = function (destination, options, events) {
     events = setupEvents(events);
     var result;
     try {
-        result = require('../../lib/create')
-        .update(destination, options, events)
-        .then(function (destination) {
+        result = require('../../lib/create').update(destination, options, events).then(function (destination) {
             var PlatformApi = require(path.resolve(destination, 'cordova/Api'));
             return new PlatformApi('android', destination, events);
         });
-    }
-    catch (e) {
-        events.emit('error','updatePlatform is not callable from the android project API, you will need to do this manually.');
-        throw(e);
+    } catch (e) {
+        events.emit('error', 'updatePlatform is not callable from the android project API, you will need to do this manually.');
+        throw (e);
     }
     return result;
 };
@@ -226,40 +218,36 @@ Api.prototype.addPlugin = function (plugin, installOptions) {
         installOptions.variables.PACKAGE_NAME = project.getPackageName();
     }
 
-    if(this.android_studio === true) {
-      installOptions.android_studio = true;
+    if (this.android_studio === true) {
+        installOptions.android_studio = true;
     }
 
-    return Q()
-       .then(function () {
-            //CB-11964: Do a clean when installing the plugin code to get around
-            //the Gradle bug introduced by the Android Gradle Plugin Version 2.2
-            //TODO: Delete when the next version of Android Gradle plugin comes out
+    return Q().then(function () {
+        // CB-11964: Do a clean when installing the plugin code to get around
+        // the Gradle bug introduced by the Android Gradle Plugin Version 2.2
+        // TODO: Delete when the next version of Android Gradle plugin comes out
 
-           // Since clean doesn't just clean the build, it also wipes out www, we need
-           // to pass additional options.
+        // Since clean doesn't just clean the build, it also wipes out www, we need
+        // to pass additional options.
 
-           // Do some basic argument parsing
-            var opts = {};
+        // Do some basic argument parsing
+        var opts = {};
 
-             // Skip cleaning prepared files when not invoking via cordova CLI.
-            opts.noPrepare = true;
+        // Skip cleaning prepared files when not invoking via cordova CLI.
+        opts.noPrepare = true;
 
-            if(!AndroidStudio.isAndroidStudioProject(self.root) && !project.isClean()) {
-              return self.clean(opts);
-            }
-        })
-       .then(function () {
-            return PluginManager.get(self.platform, self.locations, project)
-                .addPlugin(plugin, installOptions);
-        })
-      .then(function () {
-            if (plugin.getFrameworks(this.platform).length === 0) return;
+        if (!AndroidStudio.isAndroidStudioProject(self.root) && !project.isClean()) {
+            return self.clean(opts);
+        }
+    }).then(function () {
+        return PluginManager.get(self.platform, self.locations, project).addPlugin(plugin, installOptions);
+    }).then(function () {
+        if (plugin.getFrameworks(this.platform).length === 0) return;
 
-            selfEvents.emit('verbose', 'Updating build files since android plugin contained <framework>');
-            require('./lib/builders/builders').getBuilder('gradle').prepBuildFiles();
-        }.bind(this))
-       // CB-11022 Return truthy value to prevent running prepare after
+        selfEvents.emit('verbose', 'Updating build files since android plugin contained <framework>');
+        require('./lib/builders/builders').getBuilder('gradle').prepBuildFiles();
+    }.bind(this))
+        // CB-11022 Return truthy value to prevent running prepare after
         .thenResolve(true);
 };
 
@@ -279,9 +267,9 @@ Api.prototype.addPlugin = function (plugin, installOptions) {
 Api.prototype.removePlugin = function (plugin, uninstallOptions) {
     var project = AndroidProject.getProjectFile(this.root);
 
-    if(uninstallOptions && uninstallOptions.usePlatformWww === true && this.android_studio === true) {
-      uninstallOptions.usePlatformWww = false;
-      uninstallOptions.android_studio = true;
+    if (uninstallOptions && uninstallOptions.usePlatformWww === true && this.android_studio === true) {
+        uninstallOptions.usePlatformWww = false;
+        uninstallOptions.android_studio = true;
     }
 
     return PluginManager.get(this.platform, this.locations, project)
@@ -343,11 +331,9 @@ Api.prototype.removePlugin = function (plugin, uninstallOptions) {
  */
 Api.prototype.build = function (buildOptions) {
     var self = this;
-    return require('./lib/check_reqs').run()
-    .then(function () {
+    return require('./lib/check_reqs').run().then(function () {
         return require('./lib/build').run.call(self, buildOptions);
-    })
-    .then(function (buildResults) {
+    }).then(function (buildResults) {
         // Cast build result to array of build artifacts
         return buildResults.apkPaths.map(function (apkPath) {
             return {
@@ -372,10 +358,9 @@ Api.prototype.build = function (buildOptions) {
  * @return {Promise} A promise either fulfilled if package was built and ran
  *   successfully, or rejected with CordovaError.
  */
-Api.prototype.run = function(runOptions) {
+Api.prototype.run = function (runOptions) {
     var self = this;
-    return require('./lib/check_reqs').run()
-    .then(function () {
+    return require('./lib/check_reqs').run().then(function () {
         return require('./lib/run').run.call(self, runOptions);
     });
 };
@@ -387,18 +372,14 @@ Api.prototype.run = function(runOptions) {
  * @return  {Promise}  Return a promise either fulfilled, or rejected with
  *   CordovaError.
  */
-Api.prototype.clean = function(cleanOptions) {
+Api.prototype.clean = function (cleanOptions) {
     var self = this;
-    return require('./lib/check_reqs').run()
-      .then(function () {
-          return require('./lib/build').runClean.call(self, cleanOptions);
-      })
-      .then(function () {
-          return require('./lib/prepare').clean.call(self, cleanOptions);
-      });
+    return require('./lib/check_reqs').run().then(function () {
+        return require('./lib/build').runClean.call(self, cleanOptions);
+    }).then(function () {
+        return require('./lib/prepare').clean.call(self, cleanOptions);
+    });
 };
-
-
 
 /**
  * Performs a requirements check for current platform. Each platform defines its
@@ -408,7 +389,7 @@ Api.prototype.clean = function(cleanOptions) {
  * @return  {Promise<Requirement[]>}  Promise, resolved with set of Requirement
  *   objects for current platform.
  */
-Api.prototype.requirements = function() {
+Api.prototype.requirements = function () {
     return require('./lib/check_reqs').check_all();
 };
 
