@@ -25,11 +25,11 @@ var CordovaError = require('cordova-common').CordovaError;
 
 var Adb = {};
 
-function isDevice(line) {
+function isDevice (line) {
     return line.match(/\w+\tdevice/) && !line.match(/emulator/);
 }
 
-function isEmulator(line) {
+function isEmulator (line) {
     return line.match(/device/) && line.match(/emulator/);
 }
 
@@ -44,8 +44,7 @@ function isEmulator(line) {
  *   devices/emulators
  */
 Adb.devices = function (opts) {
-    return spawn('adb', ['devices'], {cwd: os.tmpdir()})
-    .then(function(output) {
+    return spawn('adb', ['devices'], {cwd: os.tmpdir()}).then(function (output) {
         return output.split('\n').filter(function (line) {
             // Filter out either real devices or emulators, depending on options
             return (line && opts && opts.emulators) ? isEmulator(line) : isDevice(line);
@@ -59,8 +58,7 @@ Adb.install = function (target, packagePath, opts) {
     events.emit('verbose', 'Installing apk ' + packagePath + ' on target ' + target + '...');
     var args = ['-s', target, 'install'];
     if (opts && opts.replace) args.push('-r');
-    return spawn('adb', args.concat(packagePath), {cwd: os.tmpdir()})
-    .then(function(output) {
+    return spawn('adb', args.concat(packagePath), {cwd: os.tmpdir()}).then(function (output) {
         // 'adb install' seems to always returns no error, even if installation fails
         // so we catching output to detect installation failure
         if (output.match(/Failure/)) {
@@ -86,8 +84,7 @@ Adb.shell = function (target, shellCommand) {
     events.emit('verbose', 'Running adb shell command "' + shellCommand + '" on target ' + target + '...');
     var args = ['-s', target, 'shell'];
     shellCommand = shellCommand.split(/\s+/);
-    return spawn('adb', args.concat(shellCommand), {cwd: os.tmpdir()})
-    .catch(function (output) {
+    return spawn('adb', args.concat(shellCommand), {cwd: os.tmpdir()}).catch(function (output) {
         return Q.reject(new CordovaError('Failed to execute shell command "' +
             shellCommand + '"" on device: ' + output));
     });
@@ -95,8 +92,7 @@ Adb.shell = function (target, shellCommand) {
 
 Adb.start = function (target, activityName) {
     events.emit('verbose', 'Starting application "' + activityName + '" on target ' + target + '...');
-    return Adb.shell(target, 'am start -W -a android.intent.action.MAIN -n' + activityName)
-    .catch(function (output) {
+    return Adb.shell(target, 'am start -W -a android.intent.action.MAIN -n' + activityName).catch(function (output) {
         return Q.reject(new CordovaError('Failed to start application "' +
             activityName + '"" on device: ' + output));
     });
