@@ -19,12 +19,12 @@
 
 var fs = require('fs');
 var et = require('elementtree');
-var xml= require('cordova-common').xmlHelpers;
+var xml = require('cordova-common').xmlHelpers;
 
 var DEFAULT_ORIENTATION = 'default';
 
 /** Wraps an AndroidManifest file */
-function AndroidManifest(path) {
+function AndroidManifest (path) {
     this.path = path;
     this.doc = xml.parseElementtreeSync(path);
     if (this.doc.getroot().tag !== 'manifest') {
@@ -32,38 +32,38 @@ function AndroidManifest(path) {
     }
 }
 
-AndroidManifest.prototype.getVersionName = function() {
+AndroidManifest.prototype.getVersionName = function () {
     return this.doc.getroot().attrib['android:versionName'];
 };
 
-AndroidManifest.prototype.setVersionName = function(versionName) {
+AndroidManifest.prototype.setVersionName = function (versionName) {
     this.doc.getroot().attrib['android:versionName'] = versionName;
     return this;
 };
 
-AndroidManifest.prototype.getVersionCode = function() {
+AndroidManifest.prototype.getVersionCode = function () {
     return this.doc.getroot().attrib['android:versionCode'];
 };
 
-AndroidManifest.prototype.setVersionCode = function(versionCode) {
+AndroidManifest.prototype.setVersionCode = function (versionCode) {
     this.doc.getroot().attrib['android:versionCode'] = versionCode;
     return this;
 };
 
-AndroidManifest.prototype.getPackageId = function() {
-    /*jshint -W069 */
+AndroidManifest.prototype.getPackageId = function () {
+    /* jshint -W069 */
     return this.doc.getroot().attrib['package'];
-    /*jshint +W069 */
+    /* jshint +W069 */
 };
 
-AndroidManifest.prototype.setPackageId = function(pkgId) {
-    /*jshint -W069 */
+AndroidManifest.prototype.setPackageId = function (pkgId) {
+    /* jshint -W069 */
     this.doc.getroot().attrib['package'] = pkgId;
-    /*jshint +W069 */
+    /* jshint +W069 */
     return this;
 };
 
-AndroidManifest.prototype.getActivity = function() {
+AndroidManifest.prototype.getActivity = function () {
     var activity = this.doc.getroot().find('./application/activity');
     return {
         getName: function () {
@@ -102,17 +102,16 @@ AndroidManifest.prototype.getActivity = function() {
     };
 };
 
-['minSdkVersion', 'maxSdkVersion', 'targetSdkVersion']
-.forEach(function(sdkPrefName) {
+['minSdkVersion', 'maxSdkVersion', 'targetSdkVersion'].forEach(function (sdkPrefName) {
     // Copy variable reference to avoid closure issues
     var prefName = sdkPrefName;
 
-    AndroidManifest.prototype['get' + capitalize(prefName)] = function() {
+    AndroidManifest.prototype['get' + capitalize(prefName)] = function () {
         var usesSdk = this.doc.getroot().find('./uses-sdk');
         return usesSdk && usesSdk.attrib['android:' + prefName];
     };
 
-    AndroidManifest.prototype['set' + capitalize(prefName)] = function(prefValue) {
+    AndroidManifest.prototype['set' + capitalize(prefName)] = function (prefValue) {
         var usesSdk = this.doc.getroot().find('./uses-sdk');
 
         if (!usesSdk && prefValue) { // if there is no required uses-sdk element, we should create it first
@@ -128,11 +127,11 @@ AndroidManifest.prototype.getActivity = function() {
     };
 });
 
-AndroidManifest.prototype.getDebuggable = function() {
+AndroidManifest.prototype.getDebuggable = function () {
     return this.doc.getroot().find('./application').attrib['android:debuggable'] === 'true';
 };
 
-AndroidManifest.prototype.setDebuggable = function(value) {
+AndroidManifest.prototype.setDebuggable = function (value) {
     var application = this.doc.getroot().find('./application');
     if (value) {
         application.attrib['android:debuggable'] = 'true';
@@ -150,7 +149,7 @@ AndroidManifest.prototype.setDebuggable = function(value) {
  * @param   {String}  [destPath]  File to write manifest to. If omitted,
  *   manifest will be written to file it has been read from.
  */
-AndroidManifest.prototype.write = function(destPath) {
+AndroidManifest.prototype.write = function (destPath) {
     fs.writeFileSync(destPath || this.path, this.doc.write({indent: 4}), 'utf-8');
 };
 
