@@ -17,8 +17,8 @@
        under the License.
 */
 
-var Q = require('q'),
-    superspawn = require('cordova-common').superspawn;
+var Q = require('q');
+var superspawn = require('cordova-common').superspawn;
 
 var suffix_number_regex = /(\d+)$/;
 // Used for sorting Android targets, example strings to sort:
@@ -29,7 +29,7 @@ var suffix_number_regex = /(\d+)$/;
 // The idea is to sort based on largest "suffix" number - meaning the bigger
 // the number at the end, the more recent the target, the closer to the
 // start of the array.
-function sort_by_largest_numerical_suffix(a, b) {
+function sort_by_largest_numerical_suffix (a, b) {
     var suffix_a = a.match(suffix_number_regex);
     var suffix_b = b.match(suffix_number_regex);
     if (suffix_a && suffix_b) {
@@ -43,9 +43,8 @@ function sort_by_largest_numerical_suffix(a, b) {
     }
 }
 
-module.exports.print_newest_available_sdk_target = function() {
-    return module.exports.list_targets()
-    .then(function(targets) {
+module.exports.print_newest_available_sdk_target = function () {
+    return module.exports.list_targets().then(function (targets) {
         targets.sort(sort_by_largest_numerical_suffix);
         console.log(targets[0]);
     });
@@ -66,38 +65,34 @@ module.exports.version_string_to_api_level = {
     '7.1.1': 25
 };
 
-function parse_targets(output) {
+function parse_targets (output) {
     var target_out = output.split('\n');
     var targets = [];
     for (var i = target_out.length - 1; i >= 0; i--) {
-        if(target_out[i].match(/id:/)) { // if "id:" is in the line...
-            targets.push(target_out[i].match(/"(.+)"/)[1]); //.. match whatever is in quotes.
+        if (target_out[i].match(/id:/)) { // if "id:" is in the line...
+            targets.push(target_out[i].match(/"(.+)"/)[1]); // .. match whatever is in quotes.
         }
     }
     return targets;
 }
 
-module.exports.list_targets_with_android = function() {
-    return superspawn.spawn('android', ['list', 'target'])
-    .then(parse_targets);
+module.exports.list_targets_with_android = function () {
+    return superspawn.spawn('android', ['list', 'target']).then(parse_targets);
 };
 
-module.exports.list_targets_with_avdmanager = function() {
-    return superspawn.spawn('avdmanager', ['list', 'target'])
-    .then(parse_targets);
+module.exports.list_targets_with_avdmanager = function () {
+    return superspawn.spawn('avdmanager', ['list', 'target']).then(parse_targets);
 };
 
-module.exports.list_targets = function() {
-    return module.exports.list_targets_with_avdmanager()
-    .catch(function(err) {
+module.exports.list_targets = function () {
+    return module.exports.list_targets_with_avdmanager().catch(function (err) {
         // If there's an error, like avdmanager could not be found, we can try
         // as a last resort, to run `android`, in case this is a super old
         // SDK installation.
-        if (err && (err.code == 'ENOENT' || (err.stderr && err.stderr.match(/not recognized/)))) {
+        if (err && (err.code === 'ENOENT' || (err.stderr && err.stderr.match(/not recognized/)))) {
             return module.exports.list_targets_with_android();
         } else throw err;
-    })
-    .then(function(targets) {
+    }).then(function (targets) {
         if (targets.length === 0) {
             return Q.reject(new Error('No android targets (SDKs) installed!'));
         }
