@@ -16,6 +16,7 @@
        specific language governing permissions and limitations
        under the License.
 */
+/* eslint no-unused-vars: 0 */
 
 var Q = require('q');
 var fs = require('fs');
@@ -42,30 +43,29 @@ function AntBuilder (projectRoot) {
 
 util.inherits(AntBuilder, GenericBuilder);
 
-AntBuilder.prototype.getArgs = function(cmd, opts) {
+AntBuilder.prototype.getArgs = function (cmd, opts) {
     var args = [cmd, '-f', path.join(this.root, 'build.xml')];
     // custom_rules.xml is required for incremental builds.
     if (hasCustomRules(this.root)) {
         args.push('-Dout.dir=ant-build', '-Dgen.absolute.dir=ant-gen');
     }
-    if(opts.packageInfo) {
+    if (opts.packageInfo) {
         args.push('-propertyfile=' + path.join(this.root, opts.buildType + SIGNING_PROPERTIES));
     }
     return args;
 };
 
-AntBuilder.prototype.prepEnv = function(opts) {
+AntBuilder.prototype.prepEnv = function (opts) {
     var self = this;
-    return check_reqs.check_ant()
-    .then(function() {
+    return check_reqs.check_ant().then(function () {
         // Copy in build.xml on each build so that:
         // A) we don't require the Android SDK at project creation time, and
         // B) we always use the SDK's latest version of it.
-        /*jshint -W069 */
+        /* jshint -W069 */
         var sdkDir = process.env['ANDROID_HOME'];
-        /*jshint +W069 */
+        /* jshint +W069 */
         var buildTemplate = fs.readFileSync(path.join(sdkDir, 'tools', 'lib', 'build.template'), 'utf8');
-        function writeBuildXml(projectPath) {
+        function writeBuildXml (projectPath) {
             var newData = buildTemplate.replace('PROJECT_NAME', self.extractRealProjectNameFromManifest());
             fs.writeFileSync(path.join(projectPath, 'build.xml'), newData);
             if (!fs.existsSync(path.join(projectPath, 'local.properties'))) {
@@ -86,7 +86,7 @@ AntBuilder.prototype.prepEnv = function(opts) {
         var propertiesFilePath = path.join(self.root, propertiesFile);
         if (opts.packageInfo) {
             fs.writeFileSync(propertiesFilePath, TEMPLATE + opts.packageInfo.toProperties());
-        } else if(isAutoGenerated(propertiesFilePath)) {
+        } else if (isAutoGenerated(propertiesFilePath)) {
             shell.rm('-f', propertiesFilePath);
         }
     });
@@ -96,7 +96,7 @@ AntBuilder.prototype.prepEnv = function(opts) {
  * Builds the project with ant.
  * Returns a promise.
  */
-AntBuilder.prototype.build = function(opts) {
+AntBuilder.prototype.build = function (opts) {
     // Without our custom_rules.xml, we need to clean before building.
     var ret = Q();
     if (!hasCustomRules(this.root)) {
@@ -104,11 +104,10 @@ AntBuilder.prototype.build = function(opts) {
         ret = this.clean(opts);
     }
 
-    var args = this.getArgs(opts.buildType == 'debug' ? 'debug' : 'release', opts);
-    return check_reqs.check_ant()
-    .then(function() {
+    var args = this.getArgs(opts.buildType === 'debug' ? 'debug' : 'release', opts);
+    return check_reqs.check_ant().then(function () {
         return spawn('ant', args, {stdio: 'pipe'});
-    }).progress(function (stdio){
+    }).progress(function (stdio) {
         if (stdio.stderr) {
             process.stderr.write(stdio.stderr);
         } else {
@@ -116,7 +115,7 @@ AntBuilder.prototype.build = function(opts) {
         }
     }).catch(function (error) {
         if (error.toString().indexOf('Unable to resolve project target') >= 0) {
-            return check_reqs.check_android_target(error).then(function() {
+            return check_reqs.check_android_target(error).then(function () {
                 // If due to some odd reason - check_android_target succeeds
                 // we should still fail here.
                 return Q.reject(error);
@@ -126,19 +125,17 @@ AntBuilder.prototype.build = function(opts) {
     });
 };
 
-AntBuilder.prototype.clean = function(opts) {
+AntBuilder.prototype.clean = function (opts) {
     var args = this.getArgs('clean', opts);
     var self = this;
-    return check_reqs.check_ant()
-    .then(function() {
+    return check_reqs.check_ant().then(function () {
         return spawn('ant', args, {stdio: 'inherit'});
-    })
-    .then(function () {
+    }).then(function () {
         shell.rm('-rf', path.join(self.root, 'out'));
 
-        ['debug', 'release'].forEach(function(config) {
+        ['debug', 'release'].forEach(function (config) {
             var propertiesFilePath = path.join(self.root, config + SIGNING_PROPERTIES);
-            if(isAutoGenerated(propertiesFilePath)){
+            if (isAutoGenerated(propertiesFilePath)) {
                 shell.rm('-f', propertiesFilePath);
             }
         });
@@ -147,10 +144,10 @@ AntBuilder.prototype.clean = function(opts) {
 
 module.exports = AntBuilder;
 
-function hasCustomRules(projectRoot) {
+function hasCustomRules (projectRoot) {
     return fs.existsSync(path.join(projectRoot, 'custom_rules.xml'));
 }
 
-function isAutoGenerated(file) {
+function isAutoGenerated (file) {
     return fs.existsSync(file) && fs.readFileSync(file, 'utf8').indexOf(MARKER) > 0;
 }
