@@ -67,6 +67,8 @@
 #define __CORDOVA_4_3_0 40300
 #define __CORDOVA_4_3_1 40301
 #define __CORDOVA_4_4_0 40400
+#define __CORDOVA_4_5_0 40500
+#define __CORDOVA_4_5_3 40503
 /* coho:next-version,insert-before */
 #define __CORDOVA_NA 99999      /* not available */
 
@@ -79,7 +81,7 @@
  */
 #ifndef CORDOVA_VERSION_MIN_REQUIRED
     /* coho:next-version-min-required,replace-after */
-    #define CORDOVA_VERSION_MIN_REQUIRED __CORDOVA_4_4_0
+    #define CORDOVA_VERSION_MIN_REQUIRED __CORDOVA_4_5_3
 #endif
 
 /*
