@@ -21,8 +21,11 @@
 
 module.exports = {
     id: 'ios',
-    bootstrap: function() {
+    bootstrap: function () {
+        // Attach the console polyfill that is iOS-only to window.console
+        // see the file under plugin/ios/console.js
+        require('cordova/modulemapper').clobbers('cordova/plugin/ios/console', 'window.console');
+
         require('cordova/channel').onNativeReady.fire();
     }
 };
-
